doc: update design documents to v0.4.0 release (#14)
With the v0.4.0 release, we now have additional support for STUN and we
dropped the `./Workspace` directory as non-practical. This diff updates
the design documents accordingly.

With respect to `./Workspace`, in particular, acknowledge that `dd-006`
is not the right place where to document the interaction with generator
scripts and installation, and defer this matter to another design
document instead.